moved HuaweiHagridSecretProvider into HuaweiHagridWspLib.kt
